Placed the Android Manifest file here
I've put in all the possible conditions we may need to include. I'm not completely sure if it is correct. Please update as you feel necessary. I have also included part of the code for the dropbox sync API but I am still working on that. 

P.S. I'm new to git..so if I'm making any mistakes please let me know.